chore: add new virtual document provider and delte previous implementation
